Checkpoint - merged realignment into refactored code - removed mspName and citizen-stub-status citizen-status is in msp-core
