--- src/app/components/msp/application/personal-info/personal-details/personal-details.component.scss
@@ -1,13 +1,0 @@
-.msp-has-error {
-  label {
-    color: #a94442
-  }
-
-  // button {
-  //   border: 1px solid #a94442;
-  // }
-}
-
-h2 {
-  color: #000000;
-}
