--- src/app/components/msp/common/citizen-sub-status/citizen-sub-status.component.html
@@ -1,3 +1,0 @@
-<p>
-  citizen-sub-status works!
-</p>
