--- src/app/components/msp/common/citizen-sub-status/citizen-sub-status.component.ts
@@ -1,12 +1,0 @@
-import { Component, OnInit } from '@angular/core';
-
-@Component({
-  selector: 'msp-citizen-sub-status',
-  templateUrl: './citizen-sub-status.component.html',
-  styleUrls: ['./citizen-sub-status.component.scss']
-})
-export class CitizenSubStatusComponent implements OnInit {
-  constructor() {}
-
-  ngOnInit() {}
-}
