--- src/app/components/msp/common/name/name.component.html
@@ -1,1 +1,0 @@
-TODO - Refactor to use <pre><common-name></common-name></pre>
